Feat: add unread style
